# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.TempCart:
    properties:
      product_varian:
        $ref: '#/definitions/entity.Varian'
      temp_cart_id:
        type: string
      temp_cart_order_id:
        type: string
      temp_cart_product_varian_id:
        type: string
      temp_cart_qty:
        type: integer
      temp_cart_user_id:
        type: string
    type: object
...
  entity.Varian:
    properties:
      product_id:
        type: string
      product_name:
        type: string
      product_varian_id:
        type: string
      product_varian_price:
        type: integer
      product_varian_qty_booth:
        type: integer
      product_varian_qty_left:
        type: integer
      product_varian_qty_warehouse:
        type: integer
      varian_id:
        type: string
      varian_name:
        type: string
    type: object
...
  handler.WebResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
...
host: 127.0.0.1:3000
info:
  contact:
    email: support@ecanteen.com
    name: API Support
  description: API documentation for E-Canteen Cashier System - A comprehensive canteen
    management solution
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Canteen Cashier API
  version: 1.2.0
paths:
...
  /tempcart:
    post:
      consumes:
      - application/json
      description: Add product variant to temporary cart
      parameters:
      - description: Cart item
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.TempCart'
      produces:
      - application/json
      responses:
        "201":
          description: Item added to cart
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.TempCart'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /tempcart/{productVarianId}/{userId}:
    delete:
      description: Remove product variant from temporary cart
      parameters:
      - description: Product Variant ID
        in: path
        name: productVarianId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item removed from cart
          schema:
            $ref: '#/definitions/handler.WebResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update quantity of item in temporary cart
      parameters:
      - description: Product Variant ID
        in: path
        name: productVarianId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart item updated
          schema:
            $ref: '#/definitions/handler.WebResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
  /tempcart/{userId}:
    get:
      description: Get all temporary cart items for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cart items
          schema:
            allOf:
            - $ref: '#/definitions/handler.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.TempCart'
                  type: array
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handler.WebResponse'
      security:
      - BearerAuth: []
      summary: Get cart items by user ID
      tags:
      - Cart
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"